app/requirements.txt
fastapi
uvicorn
python-telegram-bot==21.4
yt-dlp
httpx
pydantic
